GEOMESA-952 Pluggable BlobStores, GeoMesa BlobStore Refactoring
* BlobStore pluggable with both DataStore and BlobStore interfaces.
* renamed blobstore-core to blobstore-api, created blobstore-accumulo module
* made Blob POJO
* wrote docs for BlobStore, added EXIF tutorial
