+++ docs/tutorials/index.rst
@@ -22,6 +22,8 @@ Data In/Out
     geomesa-examples-avro
     geomesa-raster
     geomesa-examples-transformations
+    geomesa-blobstore
+    geomesa-blobstore-exif
 
 Data Analysis
 -------------
