+++ geomesa-blobstore/geomesa-blobstore-accumulo/src/test/resources/testFile2.txt
