+++ geomesa-blobstore/geomesa-blobstore-api/src/main/resources/META-INF/services/org.locationtech.geomesa.blob.api.FileHandler
@@ -1,0 +1,1 @@
+org.locationtech.geomesa.blob.api.handlers.WKTFileHandler
