--- geomesa-blobstore/geomesa-blobstore-core/src/main/java/org/locationtech/geomesa/blob/core/impl/AccumuloGeoMesaBlobStore.java
@@ -1,141 +1,0 @@
-/***********************************************************************
-* Copyright (c) 2013-2016 Commonwealth Computer Research, Inc.
-* All rights reserved. This program and the accompanying materials
-* are made available under the terms of the Apache License, Version 2.0
-* which accompanies this distribution and is available at
-* http://www.opensource.org/licenses/apache2.0.php.
-*************************************************************************/
-
-package org.locationtech.geomesa.blob.core.impl;
-
-import org.geotools.data.Query;
-import org.locationtech.geomesa.accumulo.data.AccumuloDataStore;
-import org.locationtech.geomesa.accumulo.data.AccumuloDataStoreFactory;
-import org.locationtech.geomesa.accumulo.data.AccumuloDataStoreParams;
-import org.locationtech.geomesa.blob.core.AccumuloBlobStore;
-import org.locationtech.geomesa.blob.core.GeoMesaBlobStore;
-import org.opengis.filter.Filter;
-
-import java.io.File;
-import java.io.IOException;
-import java.io.Serializable;
-import java.util.HashMap;
-import java.util.Iterator;
-import java.util.Map;
-
-public class AccumuloGeoMesaBlobStore implements GeoMesaBlobStore {
-
-    protected final AccumuloBlobStore accumuloBlobStore;
-
-    public AccumuloGeoMesaBlobStore(final Map<String, Serializable> dataStoreParams) throws IOException {
-        final AccumuloDataStore ds = genNewADS(dataStoreParams);
-        if (ds == null) {
-            throw new IOException("Error initializing AccumuloGeoMesaBlobStore");
-        } else {
-            accumuloBlobStore = new AccumuloBlobStore(ds);
-        }
-    }
-
-    public AccumuloGeoMesaBlobStore(final String instanceId,
-                                    final String tableName,
-                                    final String zookeepers,
-                                    final String user,
-                                    final String password,
-                                    final String auths,
-                                    final Boolean useMock) throws IOException {
-        this(new HashMap<String, Serializable>() {{
-            put(AccumuloDataStoreParams.instanceIdParam().key, instanceId);
-            put(AccumuloDataStoreParams.tableNameParam().key, tableName);
-            put(AccumuloDataStoreParams.zookeepersParam().key, zookeepers);
-            put(AccumuloDataStoreParams.userParam().key, user);
-            put(AccumuloDataStoreParams.passwordParam().key, password);
-            put(AccumuloDataStoreParams.authsParam().key, auths);
-            put(AccumuloDataStoreParams.mockParam().key, useMock.toString());
-        }});
-    }
-
-    private AccumuloDataStore genNewADS(Map<String, Serializable> dataStoreParams) throws IllegalArgumentException {
-        AccumuloDataStoreFactory accumuloDataStoreFactory = new AccumuloDataStoreFactory();
-        AccumuloDataStore ds = (AccumuloDataStore) accumuloDataStoreFactory.createDataStore(dataStoreParams);
-        return ds;
-    }
-
-    /**
-     * Add a File to the blobstore, relying on available FileHandlers to determine ingest
-     * @param file   File to ingest
-     * @param params Map String to String, see AccumuloBlobStore for keys
-     * @return Blob id as a string or null if put failed
-     */
-    @Override
-    public String put(File file, Map<String, String> params) {
-        return accumuloBlobStore.put(file, params);
-    }
-
-    /**
-     * @param bytes  to ingest, bypass FileHandlers to rely on client to set params
-     * @param params Map String to String, see AccumuloBlobStore for keys
-     * @return Blob id as a string or null if put failed
-     */
-    @Override
-    public String put(byte[] bytes, Map<String, String> params) {
-        return accumuloBlobStore.put(bytes, params);
-    }
-
-    /**
-     * Query BlobStore for Ids by a opengis Filter
-     *
-     * @param filter Filter used to query blobstore by
-     * @return Iterator of blob Ids that can then be downloaded via get
-     */
-    @Override
-    public Iterator<String> getIds(Filter filter) {
-        return accumuloBlobStore.getIds(filter);
-    }
-
-    /**
-     * Query BlobStore for Ids by a GeoTools Query
-     *
-     * @param query Query used to query blobstore by
-     * @return ids satisfied by the query
-     */
-    @Override
-    public Iterator<String> getIds(Query query) {
-        return accumuloBlobStore.getIds(query);
-    }
-
-    /**
-     * Fetches Blob by id
-     *
-     * @param id String feature Id of the Blob, from getIds functions
-     * @return Map.Entry<String, byte[]> map entry of filename to file bytes
-     */
-    @Override
-    public Map.Entry<String, byte[]> get(String id) {
-        return accumuloBlobStore.get(id);
-    }
-
-    /**
-     * Deletes Blob by id
-     *
-     * @param id id of the blob to delete
-     */
-    @Override
-    public void delete(String id) {
-        accumuloBlobStore.delete(id);
-    }
-
-    /**
-     * Deletes BlobStore and all stored features
-     * drops accumulo tables associated with this BlobStore
-     *
-     */
-    @Override
-    public void deleteBlobStore() {
-        accumuloBlobStore.deleteBlobStore();
-    }
-
-    @Override
-    public void close() throws IOException {
-        accumuloBlobStore.close();
-    }
-}
