--- geomesa-blobstore/geomesa-blobstore-core/src/main/resources/META-INF/services/org.locationtech.geomesa.blob.core.FileHandler
@@ -1,1 +1,0 @@
-org.locationtech.geomesa.blob.core.handlers.WKTFileHandler
