+++ geomesa-blobstore/geomesa-blobstore-handlers/geomesa-blobstore-exif-handler/pom.xml
@@ -19,15 +19,10 @@
     <name>GeoMesa Blobstore EXIF Metadata Handler</name>
 
     <dependencies>
-        <dependency>
-            <groupId>org.locationtech.geomesa</groupId>
-            <artifactId>geomesa-accumulo-datastore</artifactId>
-        </dependency>
 
         <dependency>
             <groupId>org.locationtech.geomesa</groupId>
-            <artifactId>geomesa-blobstore-core</artifactId>
-            <version>${project.version}</version>
+            <artifactId>geomesa-blobstore-api</artifactId>
         </dependency>
 
         <dependency>
