+++ geomesa-blobstore/geomesa-blobstore-handlers/geomesa-blobstore-exif-handler/src/main/resources/META-INF/services/org.locationtech.geomesa.blob.api.FileHandler
