+++ geomesa-blobstore/geomesa-blobstore-handlers/geomesa-blobstore-exif-handler/src/main/scala/org/locationtech/geomesa/blob/handlers/exif/MetadataFileHandler.scala
@@ -16,7 +16,7 @@ import java.util.Date
 import com.drew.imaging._
 import com.drew.metadata.exif.GpsDirectory
 import com.vividsolutions.jts.geom.Geometry
-import org.locationtech.geomesa.blob.core.handlers.AbstractFileHandler
+import org.locationtech.geomesa.blob.api.handlers.AbstractFileHandler
 import org.locationtech.geomesa.utils.text.WKTUtils
 
 import scala.collection.JavaConversions._
