+++ geomesa-blobstore/pom.xml
@@ -20,7 +20,8 @@
     <packaging>pom</packaging>
 
     <modules>
-        <module>geomesa-blobstore-core</module>
+        <module>geomesa-blobstore-api</module>
+        <module>geomesa-blobstore-accumulo</module>
         <module>geomesa-blobstore-handlers</module>
         <module>geomesa-blobstore-web</module>
     </modules>
